Add JxlDecoderCloseInput to decoder API
This adds a function to mark input explicitely as closed.

This is required to be used for correctness when box metadata
is gotten, because from streaming input it is not possible to
determine whether the current box is the last box, or the
current bytes of an unbounded last box are the final bytes.
Before, the decoder could return success when the user still
had more input (when requesting box content only), which is
not entirely straightforward.

It should also be used in general from this point on for
consistency, though for backwards compatibility with the
previous released decoder API versions, is not required
to be used when JXL_DEC_BOX is not used.