Add API to encode boxes, header only (no implementation)
Adds the function JxlEncoderAddBox, which can be used either before or
after encoding image frames.

Updates the JxlEncoderUseContainer documentation to noet say
"force" after all since for metadata boxes you need to call it
explicitely (the encoder cannot know on its own if boxes will be added
at the end only)

Adds documentation about the box format and individual box types so that
it is clear what types of metadata boxes this function can be used for
and which ones the encoder automatically adds.

Pair programmed with Jyrki, who also proposed to add good documentation
about the boxes and a prototype test.