Remove CPU topology detection and CPU pinning
CPU pinning by djxl/cjxl wasn't compatible with the asymmetric CPUs like
those found in most phones, since we might want to pin to the big cores
or to the small cores, depending on other factors. Newest Intel CPUs are
also asymmetric adding to this problem.

Instead, this patch removes all the CPU topology detection and pinning,
leaving it up to the OS scheduler. Defaulting num_threads to the number
of cores instead of the number of threads (which is higher with HT) has
a mixed performance impact, it may be better or worse depending on the
platform but in any case it is not too significant.

Fixes #824, fixes #734, fixes #168.