lib/jxl/encode.cc: fix ignored color encoding when XYB enabled
After the previous commits, it became possible to call AddImageFrame
without setting the color encoding provided that XYB was enabled, but
even if the color encoding was set, it was ignored if XYB was enabled.
This will fix it so it assumes SRGB if the color encoding is not set,
rather than if XYB is enabled. Entry guards to this function make it
not possible to have the color encoding be unset unless XYB is also
enabled, so this only prevents libjxl from ignoring certain calls,
rather than making calls permissible when they previously were not.