lib/jxl/encode.cc: add XYB fallback check to AddImageFrame (#839)
* AUTHORS: add Leo Izen

* lib/jxl/encode.cc: add XYB fallback check to AddImageFrame

The encoder will no longer refuse AddImageFrame if the
color encoding is set to XYB via JxlBasicInfo, even if it
is not set directly.

* lib/jxl/encode_test.cc: use proper uses_original_profile setting

This test makes sure that the color encoding is set properly before
attempting to encode, but uses_original_profile = true is the setting
to make this happen, not uses_original_profile = false. This should
fix that.

* lib/jxl/encode.cc: fix ignored color encoding when XYB enabled

After the previous commits, it became possible to call AddImageFrame
without setting the color encoding provided that XYB was enabled, but
even if the color encoding was set, it was ignored if XYB was enabled.
This will fix it so it assumes SRGB if the color encoding is not set,
rather than if XYB is enabled. Entry guards to this function make it
not possible to have the color encoding be unset unless XYB is also
enabled, so this only prevents libjxl from ignoring certain calls,
rather than making calls permissible when they previously were not.