Refector encoder for box output, and support jxlp
Various refactorings in the encoder to prepare for adding the
implementation of the add-boxes functions.

This also makes the encoder support jxlp boxes (partial codestream
boxes): it'll now add each frame in its own jxlp box, and possibly
an as early as possible jxlp box with the basic info, because it's
much simpler in the encoder to insert other boxes between frames when
using jxlp rather than a single jxlc. This is more efficient for the
decoder too since no more boxes with unbounded size are used.

Also made the decoder more strict about final jxlp being indicated:
experimenting with this new encoder feature showed that the decoder
ignored the error of not having a jxlp box marked as last.