Add JxlEncoderCloseFrames
Edited MR description after Jon's idea to keep JxlEncoderCloseInput:

Adds the separate JxlEncoderCloseFrames to close only frames.
JxlEncoderCloseInput now closes both frames and boxes.

The individual closing can also be necessary as it helps user declare
encoder can be more efficient e.g. if more frames will be added but no
more metadata boxes.

The combined JxlEncoderCloseInput is both backwards compatible and
convenient.

(Old description below:

Rename JxlEncoderCloseInput to JxlEncoderCloseFrames

This to distinguish it from CloseBoxes. The encoder needs to know
separately when each type of input ends, to allow to correctly generate
the correct box header for the final box. The CloseInput name is quite
confusing when it only closes frames and other input like boxes can
still be added.

See https://github.com/libjxl/libjxl/pull/852 for the addition of the
close boxes function.

Given that this is a rename of an always needed function, this is a
backwards incompatible change.

I believe it's still possible to do this without affecting a large
amount of usages (unlike decoder API which would require updating several
usages). The encoder API has also not been marked as frozen.

It's also OK to not submit this MR. There'll then be a function with a
bad and possibly confusing name in the encoder API in that case, but
it remains backwards compatible.)