Add use-boxes and close-boxes functions to encoder API
Add two new functions for boxes to the encoder API, to require user to
indicate they plan to add metadata boxes, and close boxes.

The reason is: it is more efficient for the encoder, if it knows no more
boxes will follow, to output a final codestream box of unbounded size.
On the other hand, if the possibility exists that the user wants to add
a box after the final codestream box, the encoder will have to output a
box with size indicated in the header, so it must hold on to the entire
output before passing it to the user to be able to write the size in the
box header.

Without the use-boxes and close-boxes functions, the encoder would
always have to operate under the assumption a box after the codestream
could follow.

In theory it's possible to have only JxlEncoderCloseBoxes, but then the
user would always have to call it, even when not intending to use any
boxes at all, so instead only assuming there can be boxes after calling
JxlEncoderUseBoxes has less such traps.