Add find module for brotli
The module looks for brotli using pkg-config if available otherwise it will look for the library in the CMAKE_PREFIX_PATH. If found a target is created that mirrors the in source build of the library.

The brotli headers do not have versioning so a minimum version is not requested.