header only: proposed encoder animation API
Adds function for setting the frame header containing animation
duration. Header-only, no implementation yet.

This change is open for discussion: let's find the best way to add this
to the encoder.

In this proposal, this setting is set to the JxlEncoderOptions, because
JxlEncoderOptions ara in reality frame options. The frame header
belongs to a frame, so setting its information to JxlEncoderOptions looks more
logical than setting it to JxlEncoder directly. Other opinions are
welcome.

JxlEncoderOptions is a confusing, JxlFrameOptions or similar could have
been a better name, you cannot use it to set boxes-related options to
the encoder for example. To clear this up, I also renamed it "frame_options"
wherever it was called "options" before and improved its documentation where used.