Apply the HLG OOTF when encoding to VarDCT and unapply it when decoding to HLG
This effectively makes VarDCT always display-referred, instead of being
display-referred if and only if the original colorspace did not happen
to be HLG. It means that the pixel data stored in the file is now always
suitable for display at a peak luminance of `intensity_target`.

JxlCmsInterface now has the responsibility of converting between scene
and display light if HLG is involved on one side and not the other. The
default implementation, JxlCms, does it and can be used for inspiration.

Images previously encoded from HLG to VarDCT with the default HLG
`intensity_target` of 1000, having not had the OOTF applied, will now be
decoded to a HLG signal brighter than the original since the OOTF will
nevertheless still be unapplied. But arguably, it was a bug of previous
encoders, since the linear signal was also too bright for display at a
peak luminance of 1000 cd/m².