use libpng for all png decoding (#923)
Currently, cjxl/djxl are using both libpng and lodepng: lodepng is used for decoding/encoding still PNGs, while libpng is used for decoding APNG.

To simplify things, we'll try to use libpng for everything. We're choosing libpng over lodepng because for distributions like Debian, libpng will already be available and is a better choice in terms of security and packaging considerations.

This PR is a first step in that direction. It makes the APNG decoding more complete: it adds support for ICC/gAMA/cHRM/sRGB, Exif/XMP metadata, and 16-bit images. This makes the code path using lodepng to decode PNG files redundant, so it is removed. This also makes it possible to correctly load APNG animations with a color profile, metadata, or 16 bits per sample (something that didn't work correctly before).

Also got rid of the ugly delete[] stuff, and added interpretation of the sBIT chunk, so we can properly encode e.g. 10-bit PNGs as 10-bit jxl files, if the PNG contains this optional chunk to indicate that only 10 bits are significant.