Remove CodecInOut::target_nits
Previously, target_nits served as a hint to set intensity_target in the
various codecs. It was set prior to decoding an image, and if non-zero,
codecs were to give it precedence over what they would have otherwise
set from the image itself (e.g. EXR would set it from the whiteLuminance
header, PQ images would get 10000 given that we have no other source of
mastering metadata than the user).

The reason for doing it this way was historical (it used to affect the
numerical pixel values that the codecs would return) and does not apply
anymore, so now we can simply let the codecs set it however they like it
and then override it if requested.