Split linear profile creation for source and destination
This fixes conversions between HLG/PQ and a non-sRGB/PQ/HLG transfer
function (e.g. gamma or a LUT profile), which would not occur during
typical use of cjxl/djxl but may occur e.g. when viewing HLG images
using the comparison tool. This now works properly although it is not
yet possible to set the peak luminance to use for the OOTF in this use
case.