Let API encoder correclty mark frames as last
The encoder was marking all frames as last.

Also fixes the encode_oneshot example to close the input, it's required
to use close input, otherwise the encoder can't know what the last frame
is. The above fix revealed this error in the example.

Documented better in the header that closing input is required.