Rename JxlEncoderFrameSettingsSetInfo and JxlEncoderFrameSettingsSetExtraChannelBlendInfo
Renames JxlEncoderFrameSettingsSetInfo to JxlEncoderSetFrameHeader, and
JxlEncoderFrameSettingsSetExtraChannelBlendInfo to
JxlEncoderSetExtraChannelBlendInfo.

This matches the decoder counterparts JxlDecoderGetFrameHeader and
JxlDecoderGetExtraChannelBlendInfo

It also makes the names, which are long and hard to remember, simpler.
THe fact that it uses the frame settings is already implied by taking a
FrameSettings argument. So overall this looks like the right thing to do
to keep the functions more memorable (shorter name and matching decoder)