Adding a general-purpose bisector tool
This allows us to do generic 'bisect to
target-size/target-bits-per-pixel' searches without having bisection
in the JXL encoder. Picking target quality settings via bisection is
relevant for performance comparison of codecs (when encoding the same
image to a given size with different encoders), but it makes more
sense to not have support for this directly in the encoder due to
potential for confusion: It in general does not make much sense to
encode images of very different visual complexity in a corpus to
identical bits-per-pixel, since this would give very unhomogeneous
quality.