libjxl implementation rely on c11 atomics (cache_aligned.cc)
It turns out that GCC requires explicitly linking to a library `atomic`
in order to support c11/atomics. gcc spec file will not handle it
directly at least not on the following Debian archs: armel, mipsel,
powerpc and riscv64.

Introduce a new cmake module to detect usage of gcc/atomic and add
missing library to the link step.

For reference:
* https://gcc.gnu.org/bugzilla/show_bug.cgi?id=104248