Fix division by zero in spline caching
Internally splines are processed in floats.
For the first segment of spline a virtual mirrored point is added.

Lets take a look at spline with control points: (16777216, 0), (16777215, 0)
Mirrored point would be (16777217, 0), but alas, it is (16777217, 0) due to floating point precision.
Next, distance between points is calculated and at some point used as a divisor...