Intensity target defaults in the C API
On the C side, leaving JxlBasicInfo::intensity_target to its initial
value of 0 lets the implementation choose a reasonable value based on
the color encoding (1000 for HLG, 10000 for PQ, kDefaultIntensityTarget
otherwise).

This also restores that behavior in cjxl, since
ConvertPackedPixelFileToCodecInOut only calls SetIntensityTarget if
ppf.info.intensity_target is 0, but starting in #1129, it was
initialized to 255.