Pfm scale factor (#1278)
* Reject PFM header where scale factor is 0.0 / -0.0

Upstream pfm specification states:

[...]
The line consists of a nonzero decimal number, not necessarily an
integer. If the number is negative, that means the PFM raster is little
endian. Otherwise, it is big endian.
[...]

ref:
* http://netpbm.sourceforge.net/doc/pfm.html

* Add a warning that PFM Scale Factor is not used

Deviate from original PFM specification and simply use Scale Factor to
check endianness.

[...]
The scale factor tells the units of the samples in the raster. You use
somehow it along with some separately understood unit information to
turn a sample value into something meaningful, such as watts per square
meter.
[...]

Ref:
* http://netpbm.sourceforge.net/doc/pfm.html