Refactor c11/atomic patch for riscv64
The commit 87fe7c1 introduced a patch to handle architectures where
c11/atomics are implemented in gcc/libatomic. It turns out that this
patch did not work for riscv64, since a symbol `__atomic_fetch_and_1`
is present in `jxl/dec_group_border.cc.o`. This eventually lead to a
compilation error:

  libjxl_dec.so.0.7.0: undefined reference to __atomic_fetch_or_1'

when compiling executable: `djxl_conformance`.

This issue was not noticed since on riscv64, the flag `-pthread`
actually imply libatomic, which is implemented by cmake line
`Threads::Threads`.