Do not attempt to create a std::string from a null pointer
Even if the API clearly document that the user input should be a
c-string (UTF-8), remove an undefined behavior from
JxlEncoderSetFrameName. The implementation does not make any difference
in between a NULL frame name and an empty frame name anyway.

[...]
Some standard library vendors, such as libstdc++, throw a
std::logic_error when a null pointer is used in the above function
calls, though not when calling std::char_traits::length(). However,
std::logic_error is not a requirement of the C++ Standard, and some
vendors (e.g., libc++ and the Microsoft Visual Studio STL) do not
implement this behavior. For portability, you should not rely on this
behavior.
[...]

Ref:
* https://wiki.sei.cmu.edu/confluence/display/cplusplus/STR51-CPP.+Do+not+attempt+to+create+a+std%3A%3Astring+from+a+null+pointer