Update AUTHORS check
This makes sure that not only the last commit in a chain of
commits is checked for authors.

At the moment this does not optimize by reducing the number
of authors by removing duplicates before running the check;
it basically runs the old check for each commit since the
common ancestor commit.