Add tests for rendering partially decoded streams with API.
The tests only document the current state of API decoding, and
show that rendering on partial input produces new details when
we have a whole group available, but does not produce new details
if we only have part of a group available.
This is the case for both lossy and lossless modes.