Broader application of JxlDecoderSetPreferredColorProfile.
This change enables JxlDecoderSetPreferredColorProfile in cases
where the image is xyb encoded and the original color encoding
is not given by an ICC profile.

Added tests for converting back to original profile and to convert
between color profiles given by enums.

Implemented the --color_space flag in djxl_ng using this new functionality.