encode: use -1 to let encoder decide level automatically
Change the codestream level mechanics in the Codestream API to
allow the encoder to automatically choose level 10 codestreams
based on the information in the basic info. This should be ABI
compatible with the previous behavior, as explicitly setting the
level to 10 should still work, and letting it default to 5 will
work in every case it did before.