Fix bit cost accounting in AuxOut.
The whole DC frame is added to kLayerDC, the whole patch reference
frame is added to kLayerDictionary, the padding bits between frames
are added to kLayerHeader, and the context map encoding is added
to the layer it was called from.

With these changes now the sum of layer bits is equal the sum of
compressed bits, and there is a warning added to benchmark stats
if it breaks again in the future.

The number and order of layers is also changed to reflect the current
state of the format.

Experiment on 31 images to see how much of bitsream is DC for some
bitrates:

```
Distance      BPP      DC%
--------------------------
d1.0        1.616    11.7%
d2.0        1.016    16.5%
d4.0        0.607    22.9%
d6.0        0.432    26.3%
d8.0        0.342    29.8%
d12.0       0.230    37.3%
d16.0       0.179    42.5%
```