Allow conversion between RGB and Grayscale in djxl.
Removed the limitation from JxlDecoderSetPreferredColorProfile
that the "grayscaleness" of the original and requested colorspaces
must match. Instead we check that the requested output format has
enough channels for color output.

DecodeImageJXL now sets the ppf.info.num_color_channels to match
the requested output color space (which is the original color space
by default), and clears ppf.info.alpha_bits if the accepted formats
can not handle an alpha channel.