Implement compressing to a specific bitrate.
This is currently rather slow (about 5x slowdown), but it is only used in the
benchmark. There is a new jxl codec param 'nr' for "normalize bitrate" and it
sets a bitrate target for each image by looking at the jpeg:95 compressed size
of the image and the target butteraugli distance.

The target bitrate it achieved by scaling the AC quantization interval lengths
by setting a custom global_scale in the quantizer. This way all other heuristic
choices are preserved.