Remove djxl's dependence on libjxl internals.
Size of djxl binary decreased from 62 MB to 3.7 MB.

All encoder functions work directly on ppf, there is no conversion to
CodecInOut, this allows djxl to be only dynamically linked to jxl library.

There are also some simplifications to the ppf structure: flipped_y parameter
was removed, now pnm encoder and decoder do the flipping on the fly, and
bitdepth_from_format was also removed since the decoders always set this
correctly for unsigned formats, and it is ignored for float formats anyway.

The basic info and ppf frame sanity checks were moved to an Encoder method
and are used by all the encoders.

Now ppf / codecio conversions only happen in tests and the benchmark.