Improve PNM codec output and ppf bit depth handling for djxl.
DecodeImageJXL will now set all bit depth info in the ppf (main,
alpha, extra channels) to match what the decoder produced, which
currently only depends on the output data type.

Setting the ppf bit depth based on command line flag is removed,
because it was never implemented in either the decoder or the
output codecs.

All encoders in extras will check that the pixel format that they
get is actually in the list of AcceptedFormats() and that the ppf
bit depth is compatible with the pixel data type (this is needed
because PPM/PGM/PAM derive the data type from the provided bit depth).

The PAM decoder and encoder can now read/write all extra channels
interleaved, since this seems to be the main purpose of the PAM
format (Portable Arbitrary Map).

EncodeImagePNM and EncodeHeader are broken up into separate
implementations for the PPM/PGM, PFM and PAM encoders.