For PQ and HLG, make it possible to generate ICC profiles that perform tone mapping
The tone mapping can be via 1D curves (smaller profile) or a 3D look-up
table (more flexible).

For PQ, the 1D version will significantly desaturate highlights whereas
the 3D version will preserve their color better.

For HLG, the 1D version assumes a system gamma of 1 (which, in a
reference environment, would be a ~300-cd/m² monitor), which is likely
to result in an image that looks quite dark on many SDR monitors or
viewing environments. The 3D version is able to perform the rendering
corresponding to a lower gamma, thereby lifting the shadows and
mid-tones, without distorting the colors as altering the 1D curves
would.

Co-authored-by: Luca Versari <veluca@google.com>