Add API functions for setting input/output bit depth.
This addresses part of the issue in #1763.

For the encoder the two choices are using the bit depth of
the pixel format (a.k.a. repeating higher bits in lower bits),
or to use the bit deth from basic info (a.k.a. zero-padding higher
bits). For the decoder there is an additional option to use a custom
bit depth.