Make cjxl output Exif/xml/jumb boxes (#1834)
* Make cjxl output Exif/xml/jumb boxes

* Add --compress_boxes option to cjxl

This allows Brotli compression of metadata boxes to be explicitly
enabled or disabled. Currently this doesn't affect boxes added
automatically for JPEG reconstruction.

* Make --compress_boxes affect recompressed JPEGs

This extends the --compress_boxes option to apply to boxes
created when a JPEG frame is added. This also flips the default to
enable compression, since that's the current behaviour for recompressed
JPEG boxes.

* Normalize Exif when reading and writing

- APNG: "Exif\0\0" is stripped on input and added on output.
- JXL: "\0\0\0\0" is stripped on input. (It's already added on output).
- None of this should touch JPEG reconstruction.

* Support non-zero offsets in Exif box