CMake: Export jxl and jxl_threads targets.
This enables downstream projects to consume libjxl through CMake's
`find_package` without relying on pkg-config or Find modules.

The only targets exported are jxl(-static) and jxl_threads(-static) in
order to match what pkg-config provides. The targets are available under
the `JPEGXL::` namespace.

When Brotli or *cms are vendored (but not bundled), they are also
installed under the `JPEGXLDeps::` namespace. Highway is excluded as it
already exports its target when vendored.

When using system dependencies, the associated Find modules are also
installed since they are required.

The tests conducted include:
- Building with and without the shared library
- Using system dependencies + bundled SKCMS
- Using vendored dependencies with LCMS2
- Using vendored dependencies and non-bundled SKCMS

For every configuration, the package was searched by setting either
`CMAKE_PREFIX_PATH` to the installed prefix or `JPEGXL_DIR` to the
location of `JPEGXLConfig.cmake`. Then an example executable was built
and linked against the exported targets.