Support cinfo reuse in jpegli encoder and decoder.
To make this work, the quantization matrices have to be defined in
the parameter setup phase, which means that we can not look at the
ICC profile, so the CICP transfer function has to be communicated
separately.

Moreover, to be more compatible with libjpeg API, we only define
one chroma quantization table by default, but we define two
chroma quant tables if the new jpegli_set_distance() API
function was called.