Change jpegli defaults to use sequential mode with fixed codes.
These are the same defaults as in libjpeg, and using different
default settings would break e.g. cjpeg linked with jpegli,
because it has only flags for -optimize and -progressive and
absent these flags it uses the library defaults.

Note that this does not change the default behaviour of
extras/enc/jpegli.cc (and thus the benchmark_xl codecs).