Check version of Apple clang correctly (#2383)
* Check version of Apple clang correctly

The `__clang_major__` macro provides a marketing version which is not
required to be consistent between vendors[1]. And indeed, Apple uses a
versioning scheme for its Clang releases that is unrelated to that of
the llvm.org releases on which they are based. The correspondences
between the versions can however be determined[2].

So, checking whether `__apple_build_version__` is defined is necessary
to determine whether an Apple version of Clang is being used, so that
the correct version check can be done. Checking the value of that macro
is simpler for the purposes of this code than checking (a different
value of) `__clang_major__`, because distinguishing between the
relevant versions of Apple Clang would also require checking
`__clang_minor__` and `__clang_patchlevel__`.

Closes: #2260

[1]: https://clang.llvm.org/docs/LanguageExtensions.html#builtin-macros
[2]: https://en.wikipedia.org/wiki/Xcode#Toolchain_versions

* fix format

---------

Co-authored-by: Moritz Firsching <firsching@google.com>