Simplify Huffman coding data structures, more consistent naming.
Define a context id for each component of each scan and gather histograms for
each context. Clustering outputs a list of Huffman tables (JHUFF_TBL can be
resued here), and a slot id for each together with a context map, then we
build coding tables for each Huffman table and use the context map and coding
tables for entropy coding. All Huffman tables and corresponding coding tables
are built in advance for all scans, and the bitstream functions for writing
DHT and SOS segments also use the context map and list of slot ids.