Fix an integer underflow bug in patch decoding.
When reading patches, the code doesn't properly check that all patches
are in bounds.

In particular, if the start of a patch becomes negative (due to the
delta coding of patch positions), but not negative *enough* to cause
`start + size` to still be negative, the decoder happily accepted the
patch. Such a patch would then lead the interval tree building code to
run into an infinite loop.