Harden current directory is a Git repository
For correctly provisioning submodules, it's not enough with checking the
current directory is a Git repository as this check may return a false
positive if the current directory is, for instance, a tarball but it's
inside a parent Git repository.

A better check is to validate whether the current path is a Git
repository and its toplevel path is equals to the absolute path of
libjxl.