Move to runtime detection of available codecs in jxl extras.
Build-time detection is problematic because depending on the
codecs available at build time libjxl_extras_codec.so can
have different set of functions exposed.

Now the codec glue-code functions are always compiled in and
avaiable to the callers, but if the corresponting codec library
is not found at build-time, these functions just return false/nullptr,
and it is the responsibility of the caller to check the return values.

This also simplifies the jxl_extras cmake file further, we do not need
the jxl_extras_codec-obj library at all.