Fix sjpeg build flag.
The -DJPEGXL_ENABLE_SJPEG=1 flag has to be added to the object library
that contains lib/extras/enc/jpg.cc otherwise it has no effect.

The build flag from djxl_main.cc is removed because it is not possible
to know when building the tool what the extras library has compiled in.