Fixes #3018 - can't save 32-bit float JXL to PNG (#3057)
In #1823 djxl got bits_per_sample argument.
It is said, that default behaviour is to export with full bit depth
of output format. But #2591 has broken this contract to save with
with depth of input format. Atfer that change it become impossible
to fulfill the contract (bits_per_sample -1 and 0 meant the same).