Fix jxl_export CMake target to be a INTERFACE library
Fix the jxl_export target to be a INTERFACE library rather than
an OBJECT library.  The latter type is only correct for libraries that
actually include any source files, and therefore produce an actual
object file.  Using it with jxl_export causes CMake to produce
an invalid ninja file.  While ninja ignores the problem, samurai is more
strict and rejects the file:

    samu: file is missing and not created by any action: 'lib/CMakeFiles/jxl_export.dir'

INTERFACE libraries "do not compile sources and do not produce library
artifacts on disk", and therefore are more correct for a header-only
target.  See:
https://cmake.org/cmake/help/latest/command/add_library.html#interface-libraries