GDK Pixbuf Loader Fixes (#3218)
* lib/jxl/decode.cc: deduplicate and correct min_size calculation

Code to calculate the proper size of the buffer exists in three places
in this function - one for the standard buffer, one for the extra
channel buffer, and one for the preview buffer. However, only the code
for the preview buffer correctly took into account the last row size
possibly being slightly smaller than the remaining rows.

This commit separates the code out into a separate function to remove
duplicate code, and uses the algorithm from the prevew buffer which was
the only correct one.

Signed-off-by: Leo Izen <leo.izen@gmail.com>

* plugins/gdk-pixbuf: consider stride when passing buffer to libjxl

This change uses gdk_pixbuf_get_pixels_with_length to get the actual
size of the buffer, with stride taken into account. The previous code
just multiplies width by height and ignores stride.

Signed-off-by: Leo Izen <leo.izen@gmail.com>

---------

Signed-off-by: Leo Izen <leo.izen@gmail.com>
Co-authored-by: Moritz Firsching <firsching@google.com>