tools/file_io.h: set binary mode for stdin and stdout on Windows
Windows stdin and stdout default to text mode, but they can be changed
to binary mode with _setmode from <io.h>. This is relevant because text
mode enables LF <-> CRLF translation, and binary mode disables it, and
it needs to be disabled for image reading/writing. This is technically
also true on Unix-like operating systems, but text and binary do the
same thing on macOS, Linux, etc.

Signed-off-by: Leo Izen <leo.izen@gmail.com>