Introduce the new jxltran tool (#4161)
* Introduce the new jxltran tool

This tool, as described in #871 will perform operations on existing jxl
files, like changing the orientation or packing a codestream in a
container.

For now, this is just a poor man `cp` for jxl files only.

* Teach jxltran how to extract JPEG XL codestream from a container

This is very bare-bone for the moment as it only supports the jxlc box,
but this is always something we can improve on in the future.