CMake: Avoid CMAKE_EXE_LINKER_FLAGS when doing link flag check
Since 88fe3fff3dc70c72405f57c69feffd9823930034, the minimum required
CMake version has been bumped to 3.16. This allows us to use the
CMAKE_REQUIRED_LINK_OPTIONS variable to set the linker flags for the
-Wl,--exclude-libs linker flag check. CMAKE_EXE_LINKER_FLAGS is not a
list, so the original patch was flaky to begin with. Unfortunately,
check_linker_flag() requires CMake 3.18, so the
check_c_source_compiles() call is still the best way for now.

This avoids configure errors on macOS environments when LDFLAGS contains
a flag that will confuse linker when a semicolon separated flag is
appended to the linker flags. This is likely to happen when a developer
has -L options in their LDFLAGS.