Avoid dllexport when built as static library on Win32
The `GenerateExportHeader` module assumes it's acting on a shared
library. When libjxl is built as a static library, we must define
`JXL*_STATIC_DEFINE` explicitly to disable symbol export.

Without this, symbols from a statically built libjxl may be
incorrectly exported as part of the ABI of any dynamic library that
links against it.

A CMake block attempting to handle this already existed, but it was
a no-op because the internal objects had already been built.